requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.10.3
groq==0.11.0
requests==2.32.3
pymupdf==1.24.1
python-docx==1.1.0
python-multipart==0.0.12



